Add splitter, yielding message bytestrings from bytestring streams
We can get blocks of bytes arbitrarily split across messages, but it's nicer to
parse complete messages together.  Attoparsec allows this with its incremental
parsing but with more overhead than a direct function on bytestrings.  The form
of this function should integrate very easily with any of the streaming
libraries, such as conduit, pipes, etc.